requirements.txt
Flask==3.1.0
requests==2.32.3
google-genai==0.8.4
